Refactored classes and added comments for readibility
