+++ WebFormsTest/Fritz.WebFormsTest.csproj
@@ -50,6 +50,7 @@
     <Reference Include="Microsoft.CSharp" />
   </ItemGroup>
   <ItemGroup>
+    <Compile Include="MasterPageReflectionExtensions.cs" />
     <Compile Include="WebApplicationProxy.cs" />
     <Compile Include="TestablePage.cs">
       <SubType>ASPXCodeBehind</SubType>
