+++ WebFormsTest/TestablePage.cs
@@ -31,6 +31,9 @@ namespace Fritz.WebFormsTest
 
     private EmptyTestServer _TestServer;
 
+    /// <summary>
+    /// The events from a Page that can be triggered within a TestablePage
+    /// </summary>
     public enum WebFormEvent
     {
       Init,
@@ -50,6 +53,9 @@ namespace Fritz.WebFormsTest
 
     }
 
+    /// <summary>
+    /// Prepare the page for testing by triggering all of the setup methods that a Page would normally have triggered when running on a web server
+    /// </summary>
     public void PrepareTests()
     {
 
@@ -190,6 +196,10 @@ namespace Fritz.WebFormsTest
 
     #endregion
 
+    /// <summary>
+    /// Simple handler to add a marker to the end of a page that shows that this is a TestablePage when debugging is enabled
+    /// </summary>
+    /// <param name="e"></param>
     protected override void OnPreRender(EventArgs e)
     {
 
@@ -200,32 +210,6 @@ namespace Fritz.WebFormsTest
         this.Controls.Add(new LiteralControl(COMMENT_MARKER));
       }
 
-    }
-
-    protected override void CreateChildControls()
-    {
-
-      if (IsInTestMode)
-      {
-        Debug.WriteLine("Controls in collection: " + Controls.Count);
-      } else
-      {
-        base.CreateChildControls();
-      }
-
-    }
-
-    public override ControlCollection Controls
-    {
-      get
-      {
-        return base.Controls;
-      }
-    }
-
-    protected internal new EventHandlerList Events
-    {
-      get { return base.Events; }
     }
 
     #region Embedded helper classes
@@ -265,6 +249,9 @@ namespace Fritz.WebFormsTest
       }
     }
 
+    /// <summary>
+    /// If there is a MasterPageFile defined, load the MasterPage and process it for this Page
+    /// </summary>
     private void CreateMasterInTest()
     {
 
@@ -309,6 +296,9 @@ namespace Fritz.WebFormsTest
       }
     }
 
+    /// <summary>
+    /// Trigger the assignment of event handlers to events if they match the prescribed naming convention
+    /// </summary>
     private void HookupAutomaticHandlersInTest()
     {
 
@@ -318,28 +308,6 @@ namespace Fritz.WebFormsTest
       mi.Invoke(this, null);
 
     }
-
-  }
-
-  internal static class MasterPageReflectionExtensions
-  {
-
-    public static void SetContentTemplates(this MasterPage master, IDictionary contentTemplateCollection)
-    {
-
-      var f = typeof(MasterPage).GetField("_contentTemplates", BindingFlags.NonPublic | BindingFlags.Instance);
-      f.SetValue(master, contentTemplateCollection);
-
-    }
-
-    public static void SetOwnerControl(this MasterPage master, TemplateControl owner)
-    {
-
-      var f = typeof(MasterPage).GetField("_ownerControl", BindingFlags.NonPublic | BindingFlags.Instance);
-      f.SetValue(master, owner);
-
-    }
-
 
   }
